+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 * added string assertions: contains, contains-not, starts-with, starts-not-with, ends-with, ends-not-with
 * added some api extensions for static usings
 * fixed multiline support for string-assertions #23
-* added enumerable-contains-assertion
+* added enumerable-assertions: contains(#24), contains-not(#25)
 
 ## New in 0.1.1
 * added support for IReadOnlyCollection